Adapted all classes for new provider URL

+++ app/src/main/java/io/pello/android/cuatrovendo/Article.java
@@ -13,8 +13,8 @@ public class Article {
     private String description;
     private Float price;
     private Date publishDate;
-    private Integer latitude;
-    private Integer longitude;
+    private Float latitude;
+    private Float longitude;
     private Integer idSeller;
     private Integer isRead;
 
@@ -70,19 +70,19 @@ public class Article {
         this.publishDate = publishDate;
     }
 
-    public Integer getLatitude() {
+    public Float getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(Integer latitude) {
+    public void setLatitude(Float latitude) {
         this.latitude = latitude;
     }
 
-    public Integer getLongitude() {
+    public Float getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(Integer longitude) {
+    public void setLongitude(Float longitude) {
         this.longitude = longitude;
     }
